+++ tests/test_circuit_ids.py
@@ -209,6 +209,23 @@ class TestCircuitNodeIds:
         )
         assert test_obj == expected
 
+    def test_intersection(self):
+        test_obj = self.ids_cls.from_tuples(_multi_index(), sort_index=False)
+        other = self.ids_cls.from_tuples([("b", 0), ("a", 3), ("a", 2)], sort_index=False)
+        expected = self.ids_cls.from_tuples([("a", 2), ("b", 0)])
+        res = test_obj.intersection(other)
+
+        # res should be sorted when inplace=False
+        assert res == expected
+        assert test_obj != expected
+
+        res = test_obj.intersection(other, inplace=True)
+        assert res is None
+
+        # test_obj index should not be sorted when inplace=True
+        assert test_obj != expected
+        assert all(expected.index == test_obj.index.sort_values())
+
     def test_sample(self):
         with patch("numpy.random.choice", return_value=np.array([0, 3])):
             tested = self.test_obj_unsorted.sample(2, inplace=False)